+++ Source/Headers/metaobjects.h
@@ -39,10 +39,10 @@ enum
 
 struct MetaObjectHeader
 {
-	u_long		cookie;						// this value should always == MO_COOKIE
+	uint32_t	cookie;						// this value should always == MO_COOKIE
 	long		refCount;					// # times this is referenced
-	u_long		type;						// object type
-	u_long		subType;					// object sub-type
+	uint32_t	type;						// object type
+	uintptr_t	subType;					// object sub-type
 	void		*data;						// pointer to meta object's specific data
 
 	struct MetaObjectHeader *parentGroup;			// illegal reference to parent group, or nil if no parent
@@ -82,11 +82,11 @@ typedef struct
 {
 	OGLSetupOutputType *setupInfo;					// materials are draw context relative, so remember which context we're using now
 
-	u_long			flags;
+	uint32_t		flags;
 	OGLColorRGBA	diffuseColor;					// rgba diffuse color
 
-	u_long			numMipmaps;						// # texture mipmaps to use
-	u_long			width,height;					// dimensions of texture
+	uint32_t		numMipmaps;						// # texture mipmaps to use
+	uint32_t		width,height;					// dimensions of texture
 	GLint			pixelSrcFormat;					// OGL format (GL_RGBA, etc.) for src pixels
 	GLint			pixelDstFormat;					// OGL format (GL_RGBA, etc.) for VRAM
 	void			*texturePixels[MO_MAX_MIPMAPS]; // ptr to texture pixels for each mipmap
@@ -207,7 +207,7 @@ typedef struct
 //-----------------------------
 
 void MO_InitHandler(void);
-MetaObjectPtr MO_CreateNewObjectOfType(u_long type, uint64_t subType, void *data);
+MetaObjectPtr MO_CreateNewObjectOfType(uint32_t type, uintptr_t subType, void *data);
 MetaObjectPtr MO_GetNewReference(MetaObjectPtr mo);
 void MO_AppendToGroup(MOGroupObject *group, MetaObjectPtr newObject);
 void MO_AttachToGroupStart(MOGroupObject *group, MetaObjectPtr newObject);
