+++ Source/Items/Items.c
@@ -34,7 +34,7 @@ extern	OGLVector3D			gDelta;
 extern	NewObjectDefinitionType	gNewObjectDefinition;
 extern	OGLBoundingBox 		gObjectGroupBBoxList[MAX_BG3D_GROUPS][MAX_OBJECTS_IN_GROUP];
 extern	OGLSetupOutputType	*gGameViewInfoPtr;
-extern	u_long				gAutoFadeStatusBits,gInfobarUpdateBits;
+extern	uint32_t				gAutoFadeStatusBits,gInfobarUpdateBits;
 extern	Boolean				gAreaCompleted;
 extern	PlayerInfoType		gPlayerInfo[];
 extern	const float gWaterHeights[NUM_TRACKS][6];
