+++ Source/Screens/ModeScreen.c
@@ -260,7 +260,7 @@ short				i;
 
 	if (FSMakeFSSpec(gDataSpec.vRefNum, gDataSpec.parID, ":images:GameModeScreen", &spec))
 		DoFatalAlert("DoGameMode: background pict not found.");
-	gBackgoundPicture = MO_CreateNewObjectOfType(MO_TYPE_PICTURE, (u_long)gGameViewInfoPtr, &spec);
+	gBackgoundPicture = MO_CreateNewObjectOfType(MO_TYPE_PICTURE, (uintptr_t) gGameViewInfoPtr, &spec);
 	if (!gBackgoundPicture)
 		DoFatalAlert("DoGameMode: MO_CreateNewObjectOfType failed");
